Update for Ookla/speedtest-cli module changes
Suggested changes to address #67. The changes regarding speedtest-cli assume the system is using Python 3.8 (ie OPNsense 21.7 or later), and that it may have an older Python 3.7 speedtest-cli module installed if the plugin was used before OPNsense 21.7. There is probably a more robust way to implement this (for example, test which Python version is being used)?